requirements.txt
flask
reportlab
google-generativeai
python-dotenv
python-docx
PyPDF2
docx2txt
